+++ admin_pages/general_settings/templates/countries_settings.template.php
@@ -1,7 +1,10 @@
 <?php
-/** @var EE_Question_Form_Input $countries */
-/** @var string $country_details_settings */
-/** @var string $country_states_settings */
+/**
+ * @var EE_Question_Form_Input $countries
+ * @var string $country_details_settings
+ * @var string $country_states_settings
+ * @var string $CNT_name_for_site
+ */
 ?>
 <div class="padding">
 
@@ -17,10 +20,23 @@
     </table>
     <br/>
     <p>
-        <?php _e(
+        <?php
+        esc_html_e(
             'The country that is selected above will populate the Country Details settings and the options for States/Provinces. This information will be used throughout Event Espresso including for registration purposes and how currency is displayed. If you make a change to the country on this page, it is important that you also update your Contact Information on the Your Organization tab.',
             'event_espresso'
-        ); ?>
+        );
+        ?>
+    </p>
+    <p id="country-currency-setting-disabled-pg" style="display: none;">
+        <span class="reminder-spn">
+            <?php printf(
+                esc_html__(
+                    'Currency setting inputs are currently only enabled for the country "%1$s" which is the selected country for the site. This can be changed in the "Contact Information" section under the "Your Organization" tab of the Event Espresso - General Settings.',
+                    'event_espresso'
+                ),
+                $CNT_name_for_site
+            ); ?>
+        </span>
     </p>
     <div id="country-details-settings-dv">
         <h2 class="ee-admin-settings-hdr">